Removed daos_metrics output that showed iteration and metrics count.
Updated d_tm_init() so that it creates a root node with the
name "srv_idx" instead of "rank".  When printing the telemetry
tree from the root node, it will now print "srv_idx <num>"
instead of "rank <num>".  This better aligns with the use
of rank as a globally unique identifier, and srv_idx as a
locally unique server index identifier.

Added a metric "server/rank" that stores the globally unique
rank ID assigned to the server instance generating the metrics.

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>