DAOS-5877 rebuild: Only merge queued rebuild tasks for sequential operations (#4056)
DAOS-5877 rebuild: Only merge queued rebuild tasks for sequential operations

Prior to this patch, rebuild will merge any tasks together that are of
the same operation. This used to always be a safe thing to do when
rebuild handled only failure recovery. However, now that rebuild handles
multiple different types of operations, it was still able to merge the
same type together, even if there was a separate operation that happened
in between. This caused the rebuild version to be set incorrectly and
placement to return incorrect data, resulting in data loss.

With this fix, rebuild will only merge tasks that are safe to merge
because the same type of operation happened after another of the same
type without any other types in between. For example, fail -> fail ->
fail will merge into just one task, but fail -> reint -> fail will leave
three independent queued tasks, which are executed in that order.
Without this patch, this would merge the two fails together and handle
them both prior to the reintegration, which was incorrect.

Signed-off-by: Byron Marohn <byron.marohn@intel.com>