DAOS-6244 IO Server: Coverity 275736 (#4046)
Reworked option parsing for the affected elements
to address the Coverity scan failures.  The initial
problem was that we were comparing the result
against a variable with the wrong size so the error
comparison never evaluated as expected.  Code has
since evolved to work around a faulty comparison.  
Moved the troubled string conversions to atoi() as is
used elsewhere, to remain compatibility with code 
that was using these values and interpreting out of 
range values on its own.  This has less impact than
reworking all of the locations where the values were
used.

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>