do not allow passing in epoch, and increase key buf size
Based on testing, passing in an epoch of an already
created snapshot does not work when passed to
daos_oit_open. Passing in only the DAOS_SNAP_OPT_OIT
flag without the DAOS_SNAP_OPT_CR to daos_create_snap_opt
does not work. It returns a snapshot not supported
error. So, the check for a passed in epoch on the copy
was removed, because it doesn't seem to be
supported at this time. Also, the key buf size
was increased due to the fact that for some test
runs 32 was not big enough, and many keys were at
a size of 68. It was increased to 512 to handle that.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>