DAOS-5928: crash_ior.py failure to stop ior command
Observed that the SIGTERM signal is not being propagated through the test
client so the IOR command is never killed in that node and after 5 secs
the parent PID in test server is killed with SIGKILL signal and the test
fails.

Modified the "def stop()" function in command_utils.py source file adding
another SIGTERM signal to the signal_list list. By doing this the
SIGTERM signal is now killing the child process and the test finish
successfully.

Test-tag: pr,-hw crashior

Signed-off-by: David Maldonado Moreno <david.maldonado.moreno@intel.com>