DAOS-6440 vos: remove invalid DTX status check (#4303)
The dtx_leader_begin() may generate DTX entry in DRAM before any
real modification. The index for such DTX entry will not be set
until all related modifications are done locally. If the DTX is
aborted before that, the dtx_rec_release() that is triggered by
DTX abort logic will hit assertaion when check the index of DTX
entry, that is wrong, and should be removed.

Signed-off-by: Fan Yong <fan.yong@intel.com>