DAOS-623 build: Remove spurious message and unify makedirs
SCons was printing a message for calls to os.makedirs that
threw an exception because the directories already exist.

Rather than throwing and catching an exception, instead
first check for existence.   And only throw an exception if
the directory cannot be made.

Also, consolidate this code into a single function

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>