DAOS-6386 iosrv: Fix D_FREE() on undefined ptr (#4425)
Missed this case in testing the clean shutdown patch. The sc_locate()
callback defined for the mgmt worker was not setting the path to
the db (because there is no mgmt db anymore), but this caused
D_FREE() to fail. Instead, just create a dummy path that won't
cause stat() to fail.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>