Update test with logging and run pool destroy in both cases.
Enhance TestPool.start_rebuild() with a boolean argument "force" to
specify whether the --force= argument is to be used with dmg system stop.
And update rebuild/pool_destroy_rebuild.py test to specify force=True when
stopping the rank to trigger rebuild.

Enhance TestPool.wait_for_rebuild() with a "smart sleep" to query the pool
periodically until a successful reply, before then monitoring for the
rebuild status to know that rebuild has started or completed.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>