DAOS-6573 container: fix container reference issue
In original implementation, when fail to open the container,
the container reference may be dropped unexpectedly, that may
caused the container to be released while others are using it.

The patch also addjusts DTX batched commit register/deregister
logic to avoid potential assertion for container status check.
It resolves the race between dtx_batched_commit_deregister and
dtx_batched_commit_register for re-open the container.

Signed-off-by: Fan Yong <fan.yong@intel.com>