DAOS-6367 object: workaround for OC_SX/OC_RP_XSF (#4433)
* DAOS-6367 object: workaround for OC_SX/OC_RP_XSF

The main issue of OC_SX/OC_RP_XSF objects is these objects
cannot be rebuilt or rebalanced. This is the nature of their
definition: spreading to ALL targets in the pool. "ALL" can
be changed over time when user including/excluding targets
from the pool, so there is no way for the placement algorithm
to generate the same layout after membership change of a pool.

This patch implements a function:
  daos_oclass_fit_max(int oc_id, int domain_nr,
		      int target_nr, int *oc_id_p)

If the input @oc_id is OC_SX or any class which requires more
targets than the pool actually has (target_nr), or the input
@oc_id is OC_XSF or any class which requires more domains
than the pool actually has (domain_nr), this function can map
the input class to another class which can fit into the pool.

It also changes DFS to call this function, so it can convert
OC_SX/OC_RP_XSF to anthoer class with exact targets number.

This patch also defines more object classes.

TODO:
- add more object classes to the class table
- add a generic function to select object class for middleware

Signed-off-by: Liang Zhen <liang.zhen@intel.com>