DAOS-6317 pool: support for admin-defined domain in pmap
Some changes are required to the pool map code to support variable
number of fault domain.

This patch implements the following changes:
- take a level (starting at 0 from the leaf = target) for the
container property.
- target and node levels are hardcoded to respectively 0 and 1.
- 253 levels above nodes are then available to the user. The
control plane will maintain the association level ID to
name reported in the yaml file.
- keep the rack level (= 2) defined in the source code for test
purpose, but the admin can reuse this level for something else.
- change the component ID (= level) to 8 bits (instead of 16) for
a total of 256 which is way more than enough.

Signed-off-by: Johann Lombardi <johann.lombardi@intel.com>