DAOS-6722 placement: placement algorithm improvements
The issue with current implementation is it spend a lot of time to
find unused target when try to widely place. It calculate new hash
in the loop until it will point to next unused target. The time of
this calculation is unpredictable and in worst case can take a very
long time.

The idea of improvement is to make the time of placement predictable
and independent from widely placing. Now if first hash points to
used target it try to find first unused target around proposed in
the window of 64 targets size. If it still fails (all 64 targets
are used in this window) it search unused target in whole available
targets array. So, only one hash calculation is required now and in
worst case two linear searches into bits arrays.

Additionally it searches for root domain only once and then pass it
into subsequent functions.

Same of structure members were rearranged for better cache utilization
and prefetching.

The debug dumps functions are called only if appropriate debug level
is set.

Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>