DAOS-6723 iosrv: Call sm_fini when shutting down server modules.
This call must have been dropped at some point in the past and
was causing a number of server memory leaks.
Fix a crash caused by double-shutdown of the rdb code.
Add D_ALLOC logging to one place in the drpc code.

Overall this improves the shutdown, and closes a number of
memory leaks in the server.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>