DAOS-5434: weekly-test: ObjOpenBadParam: test_invalid_handle and test_closed_container_handle: Test died without reporting the status
Tests 'test_invalid_container_handle' and 'test_closed_container_handle' were
failing not reporting the expected status because in the obj_layout_create()
function in object/cli_obj.c L308, when trying to get the pool from an
incorrect container handler, the test was aborting as soon as it noticed
that pool==NULL.
In order to fix this the D_ASSERT call was replaced with a D_GOTO call
instead, passing the expected '-DER_NO_HDL' return code. With this, the
'open()' function of the DaosObj class can raise the DaosApiError to the
test itself.

Test-tag: pr,-hw object

Signed-off-by: David Maldonado Moreno <david.maldonado.moreno@intel.com>