DAOS-6745 Systemd: Make systemd unit files conditionally install based on version
Some of the systemd unit file directives used change their name between the
versions of systemd. This means depending on which platform you build/install
DAOS on the unit files will be incorrect. Modify the rpm spec to selectively
install a pre systemd 230 unit file on EL7 and a post systemd 230 unit file on
everything else.

Signed-off-by: David Quigley <david.quigley@intel.com>