CART-9456 Selftest: Add statistics for M:EP analysis.
The data for all endpoints based on an individual masters are resorted
so that so that pareing of master:endpopint are made.  The statistics
for each of these are created and analyized.  The lookup for the
expected results and scalaing factors is expanded to include
master-endpoint specifications, allowing for greater detail
examination of results.

Signed-off-by: Steven Bollinger <stevenx.bollinger@intel.com>