DAOS-4891 test: Use multiple access points
Bump up the number of access points to the number of servers used or 4, whichever is lower.

Updated test.py where it sets self.access_points. It used to set just the first host in
self.hostlist_servers (daos_server host). Now it checks the number of servers and if it's 4
or less, use all servers as AP. If it's above 4, use the first 4 servers as AP.

Quick-Functional: true
Test-tag: daos_object_query server_restart server_reformat dynamic_start_stop
Signed-off-by: Makito Kano <makito.kano@intel.com>