DAOS-6713 control: Fix agent systemd unit for ofi+verbs;ofi_rxm (#4610)
The verbs provider was no longer available when running
the daos_agent via systemd.  The daos_server was capable
of finding and using the verbs provider, but the agent
was not.  This mismatch caused the agent to provide the
default connection information to the client, which was
not suitable for a valid verbs connection.

The root cause of the failure was a lack of memory
resources available to the daos_agent service.
libfabric could not initialize the queue pairs with the
default memory limits.

Removing the memory limitation for the daos_agent service
addresses the problem.

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>