DAOS-6302 placement: Fix for Layout Map Extend (#4402)
When extending the placement map to account for shards that are
being rebuilt, the group indexing logic was incomplete. Also,
comparing fseq to shard index for setting the rebuilding flag
was incorrect

In order to use daos_oclass_attr_find, which is in object.h, the
obj_class_init must be called, but it is in obj_internal.h. Moved
obj_class_init/fini to same API level as daos_oclass_attr_find so
obj_class can be initialized from tests.

Updated run_test.sh so that the jump map placement tests are run with
other unit tests. Because CI runs run_test.sh with valgrind,
several memory issues where discovered in the new jump map placement
tests and the placement
algorithm.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>