DAOS-6823 build: Move utils/sl to site_scons.
Move the code to where scons expects it to be, rather than
in utils/sl which is a legacy of having the scons_local
git submodule in the past.

Remove LICENSE, NOTICE and README.md which are all outdated now
the code is part of daos.

Update some incorrect documention links.
Change dockerfiles to new build location.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>