DAOS-6824 control: Fix pool destroy RPC handler
In some edge cases, a pool could be left in a state where the
MS entry exists but the data plane resources have been destroyed,
and subsequent attempts to destroy the pool would fail.

This commit changes the pool destroy handler to properly handle
various scenarios where the destroy process is interrupted or
fails in unexpected ways.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>