DAOS-6714 rebuild: Delete redundant objects prior to reintegration
This patch changes the rebuild migration behavior to delete individual
existing objects prior to reintegration, rather than whole containers.

This means that any object that exists on a server prior to that server
being reintegrated will be deleted only if the rebuild migration process
detects that data as existing on another target and attempts to migrate
it.

There are several practical results of this enhancement:
- Containers will no longer be deleted incorrectly on other local
  targets when reintegrating just one target (DAOS-6162)
- Objects that have no rendundancy but are still healthy on the
  reintegrating node will still be available after reintegration.
- Objects will still be deleted prior to migration, so that VOS will not
  be unhappy about migrating the same data on top of itself.

Importantly, this also means that the hash table that tracks deletions
must track object IDs. If there are millions of object IDs in a pool
that must migrate to the reintegrated rank, that hash table will contain
those millions of entries, potentially consuming a significant amount of
memory.

Skip-test: true

Signed-off-by: Byron Marohn <byron.marohn@intel.com>