DAOS-6851 test: Fixed and improved daos object query test (#4716)
The test was failing because the object class code changed.
Updated the test so that it verifies daos object query output with S1, RP_2G1, RP_3G1, and RP_2G2.
It checks oid and grp_nr values. It also checks the number of grp and replica rows in the table based
on the object class.

Quick-build: true
Skip-unit-test: true
Test-tag: daos_object_query cpu_usage
Signed-off-by: Makito Kano <makito.kano@intel.com>