DAOS-6745 build: Install correct systemd units based on version (#4593)
Some of the systemd unit file directives used change their name between the
versions of systemd. This means depending on which platform you build/install
DAOS on the unit files will be incorrect. Modify the rpm spec to selectively
install a pre systemd 230 unit file on EL7 and a post systemd 230 unit file on
everything else.

Signed-off-by: David Quigley <david.quigley@intel.com>