DAOS-6898 systemd: Add missing line to pre230 systemd unit files (#4814)
The LimitMEMLOCK=infinity line that was added to fix verbs support in
daos_agent landed after the file was copied to create the pre230 version of the
daos_agent service file. This adds the line into the pre230 version of the
daos_agent service file to get verbs working again by default on centos 7. This
change only affects the centos 7 builds.

Signed-off-by: David Quigley <david.quigley@intel.com>