DAOS-6748 placement: Do not clear used_tgts in get_target (#4751)
Do not clear used_tgts in get_targets, since we do not want
put multiple shards in a single VOS target yet, which could
happen when there are different VOS targets in each fault
domain.

Add dom_occupied to get_targets to represent the domain whose
targets has been used up for the objects, so to avoid the
domain to be choosen.

Do not need rehash within the same rank, if the targets has been
used, let's just walk through the targets to find the avaible
spots.

Signed-off-by: Di Wang <di.wang@intel.com>