DAOS-6505 pool: Limit cached VOS pool open handles
Enable the LRU cache for VOS pool so that a limited number of
VOS pools remain open.  PMDK uses a pthread_key per open pool
and POSIX has a hard limit on the number of pthread keys that
can be used at once.

PMDK has a feature #5172 that will get rid of this issue by
allowing us to configure a pool for single threaded use.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>