DAOS-6866 telemetry: APIs for creating metrics support fmt spec (#4740)
The telemetry API functions below now support a format
specifier with optional arguments, printf style, to specify
the full path and name of the metric to create or read.

    d_tm_add_metric()
    d_tm_increment_counter()
    d_tm_record_timestamp()
    d_tm_take_timer_snapshot()
    d_tm_mark_duration_start()
    d_tm_mark_duration_end()
    d_tm_set_gauge()
    d_tm_increment_gauge()
    d_tm_decrement_gauge()

It is no longer necessary to preformat the path string
as this functionality is now incorporated into the API.

It is no longer necessary to add a trailing NULL argument
after the path specifier as this is no longer required
when using a format specifier.

Modifies d_tm_increment_counter() to allow the caller to
specify an arbitrary increment value.

Modifies d_tm_mark_duration_end() so that it now takes
an 'err' input.  If non-zero, it aborts the interval
calculation so that this interval is not added to the
stats.  Used when the caller knows it has a duration
it does not want counted due to some error condition
that invalidates the duration.

Changed D_TM_SUCCESS -> DER_SUCCESS

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>