DAOS-6915 control: Don't allow new ranks to claim existing (#4833)
As a leftover from some older logic, servers with a new
UUID were being allowed to present a join request for an
existing rank, with the result that the new server replaced
the old one in the system membership. This commit removes
the ability to implicitly replace an existing server. Future
work will add support for explicit replacement via new API
surface.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>