DAOS-6319 rsvc: Optimize bootstrapping latencies
To optimize pool create latencies, when creating and bootstrapping an
rsvc with an initial membership, this patch lets one replica to start
campaigning without waiting for the election timeout. Also, the patch
replaces the 1 s constant backoff in ds_pool_svc_create with an
exponential one starting from a few milliseconds. The greatest factor of
the latency of a pool create operation will more likely be the creation
of the PMDK pools, as a result.

Signed-off-by: Li Wei <wei.g.li@intel.com>