DAOS-6989 dtx: initialize daos_unit_oid_t::id_pad_32
It must be initialized as zero, because it will be used as
part of the vos object cache index. If it is uninitialized,
then we may cannot find out former created object or others
cannot find out the object created by current udpate.

Master-PR: https://github.com/daos-stack/daos/pull/4940

Signed-off-by: Fan Yong <fan.yong@intel.com>