DAOS-6958 event: proper event handing in case of progress errors
crt_progress can sometimes return errors that are not timeout. Those
are not properly handled in synchronous IO calls and cause the private
event to be in an unusable state for following IOs.

This PR checks the error from cart and reinits the private event in
case an error was returned so it can be resused rather then returning
to the user.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>