DAOS-6943 test: Use JSON for dmg pool list
storage_ratio test uses non JSON to create a pool, but we wouldn't be able to tell the created pool UUID because we don't
do the stdout scraping. Since we need to destroy the pool at the end of the loop, we call dmg pool list to find the UUID.
pool_list uses stdout, so updated to use --json. There are several tests that are impacted by this change, so updated them.

Quick-build: true
Skip-unit-tests: true
Test-tag: storage_ratio DAOS_5610 list_pools dynamic_server_pool
Signed-off-by: Makito Kano <makito.kano@intel.com>