DAOS-1452 tests: dump ULT stacks in CI upon need
Since Argobots-1.1a1 the feature to dump all ULT stacks
is present, and its usage has also been implemented in DAOS
which can be triggered by sending SIGUSR2 to an engine
process.

This patch allows to also use this feature in CI when a
hung/timeout situation is being encountered.

Change-Id: Icacc5ebf9782b0170b70e0cc25d73c3ed6373e0b
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>