DAOS-7007 md,test: fix NO_OF_MAX_CONTAINER setting
Previous commit 0b86a68 inadvertently set this constant too high.
The constant is intended to represent a near-maximum number of
containers that can be created in a pool, from the perspective of the
pool's metadata (rdb) capacity, based on empirical testing.

That commit worked for the existing metadatafill tests, which currently
passes in testing. However it negatively impacted metadata_free_space,
and also impacts a subsequently-developed test metadata_der_nospace.

With this change the constant is set based on the actual number of
containers that can be created before the pool service would return
DER_NOSPACE.

Also with this change, a setting that is no longer applicable is
removed from metadata.yaml. And, the scm_class is now set to dcpm
to exercise persistent memory rather than tmpfs in regression testing.

Parallel-build: true
Skip-unit-tests: true
Skip-coverity-test: true
Test-tag: metadatafill metadata_free_space metadata_der_nospace

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>