DAOS-7123 dfuse: Fix race condition in dfuse.
Correctly protect against simultanously connecting to and
forgetting a container by using hash tables to track pool
and container connections rather than a simple list.

What could happen was that a forget could be issued, causing
an inode to be closed and a container to be closed, but a
different thread could have already checked the pool list and
decided to re-use the pool.  Lack of ref counting on the list
entries meant this was unsafe, so the switch is to change to
a reference counted hash table which avoids the issue and is
generally safer.

Skip-func-hw-test: true
Skip-func-test: true
Parallel-build: true

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>