DAOS-5758 pl: fixes for placement
1: Add allow_status to layout generation, so only
the target with allow status can be existed in
the layout. And remove op_type/for_reint and some
duplication to make those placement algorithm easier
to follow.

2. During reintegration reclaim process, it needs
to compare the shard id as well, i.e. if shard id
is different, the object needs to be deleted as
well.

Signed-off-by: Di Wang <di.wang@intel.com>