DAOS-7047 control: fix panic when syslog is not configured
Event logger incorrectly assumes that a syslogger for a given priority
can be created (which requires syslog to be configured on the local
host). Missing syslog capability on host triggers an exception in
daos_server. Fix by reporting error when failing to create syslogger
and skipping missing syslog priority targets when logging event.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>