DAOS-7051 ras: Improve engine init sequence
Currently, if a RAS event is raised before the engine's
rank is set, an assertion will be tripped and the engine
will crash.

This commit changes dss_self_rank() to print an error
and return CRT_NO_RANK instead of asserting. It also moves
the call to gethostname() to the beginning of server_init()
so that init_event() doesn't fail.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>