DAOS-5952 dtx: not restart TX if with specified epoch
If the TX sponsor specified epoch for the transaction,
such as for tx_open_snap() or tx_local_open(), then we
need to keep such epoch instead of restarting it even
if the pool map is stale.

On the other hand, for the fixed epoch TX, it must be
read-only transaction, and will not cache IO locally.

This patch also adjusts EC data recovery to use highest
epoch replied from server for the case that server does
not provide shadow epoch through recovery-list, such as
the CSUM error caused singlv data recovery.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Co-authored-by: Xuezhao Liu xuezhao.liu@intel.com