DAOS-5566 tests: The paths are for CaRT (mostly)
Except for a few dfuse ioil tests, `/usr/lib/daos/TESTING/tests` contains CaRT
tests.  So although the name of the directory is generic (`tests`), it seems to
be mostly a directory for CaRT tests. Also, this Jira ticket is for the removal
of relatives paths in the _*CaRT*_ test YAML files. In the case that this
subroutine becomes useful for non-CaRT tests, the code is portable enough to be
moved elsewhere (e.g., `apricot/test.py`).

Signed-off-by: Ethan Mallove <ethanx.a.mallove@intel.com>