DAOS-5758 pl: fixes for placement (#5146) (#5173)
1: Add allow_status to layout generation, so only
the target with allow status can be existed in
the layout. And remove op_type/for_reint and some
duplication to make those placement algorithm easier
to follow.

2. During reintegration reclaim process, it needs
to compare the shard id as well, i.e. if shard id
is different, the object needs to be deleted as
well.

3. In particular, fix find_reint and main placement
APIs returning too many items when the pool map
contains simultaneous reintegration, drain, and
failure operations. This is a possible real-world
scenario that would be triggered when a reintegration
is running and something fails.

4. re-enable a few placement tests and add multiple
simultaneous states tests.

Signed-off-by: Di Wang <di.wang@intel.com>
Signed-off-by: Byron Marohn <byron.marohn@intel.com>