DAOS-5566: Remove relative paths from CaRT test YAML (#4665)
Find the path to the scons-built install dir based on the location of the
avocado script (if RPM is not installed).  Add DAOS_ENV var to explicitly use
developer-built ./install dir for avocado testing.

Access check in crt_launch is broken because it doesn't set the path to the
executable using the PATH env var.  Skip the access() check, and inspect return
value of execve() instead (if execve does return).  If execve succeeds, it won't
return.  https://man7.org/linux/man-pages/man2/execve.2.html

Signed-off-by: Ethan Mallove <ethanx.a.mallove@intel.com>