DAOS-7115 test: Fix rebuild_test.py failing in CI
Due to recent increase of resources on vms the test
started failing and brought up some test code issues
as well as daos handling of server start when resources
available on nodes are limited. As a workaround number
of targets for the test is reduced and the way calculations
are done for expected values is modified a little.

Quick-Functional: true
Test-tag: rebuild_tests

Signed-off-by: Saurabh Tandan <saurabh.tandan@intel.com>