DAOS-5109 cont: increase tls dt_cont_cache key size (#5062)
Before this change, specifying the same container UUID value in
distinct daos pools results in only a single entry in a daos_engine's
thread-local storage dt_cont_cache, associated with one of the pools.
Yet all cont_child_lookup() calls searching by container uuid return
this entry, even if that lookup is in the context of another daos pool.
In daos container queries this leads to opening the wrong vos container
(associated with a different daos pool). Assertions in the daos_engine
are then observed, e.g., in ds_cont_tgt_query_handler().

With this change, the key for dt_cont_cache is expanded to be the
size of two uuid_t, a concatenation of the container uuid and the uuid
of the pool in which the container exists. The ds_cont_child structure
now includes another uuid_t for the pool, immediately following and
contiguous in memory to the preceding sc_uuid (container uuid).

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>