DAOS-6004 control: Fix system reformat
As a result of moving the Management Service up into
the Control Plane, the system reformat feature was
broken because the raft database held files open on
SCM, preventing reformat from succeeding on the MS
replicas.

This commit re-enables the feature with a slight workflow
modification. Before, the admin could run the following
commands:

  # dmg system stop (stop all engine ranks)
  # dmg storage format --reformat

Now, there is an extra step which should make the process
more explicit and also works better with the new architecture:

  # dmg system stop
  # dmg system erase
  # dmg storage format (--force may be necessary)

Future work should refine this flow to remove the need for the
--force flag to storage format in most cases by removing all
metadata in the erase step.

Master-PR: https://github.com/daos-stack/daos/pull/5061

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>