DAOS-7169 control: Fix intermittent deadlock on server startup
The work for DAOS-6004 modified EngineInstance.NotifyStorageFormat()
to accept a context, on the theory that the goroutine should be
cancelable if the parent context is canceled. Normally this is a
good idea, but in this case it was an error because on highly-loaded
systems the gRPC context from StorageFormat() could be canceled before
the starting instance is able to read from the storageReady channel,
leading to a deadlock on startup.

Also bypasses the NVMe format logic if the engine config has no
block devices. This avoids printing a confusing message about
formatting NVMe even though there's no NVMe in the config.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>