DAOS-7177 control: report nvme result from storage prepare
Update dmg storage prepare command output to display results of SCM and
NVMe prepare operations in a way that will illustrate which actions were
performed in the cases that either one or the other failed and if only
one or the other was requested. If a particular host fails NVMe prepare
but succeeds SCM prepare for example, a host error will be displayed for
the NVMe failure and the coalesced result table will contain a separate
entry for the failed host indicating SCM success but no NVMe result.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>