DAOS-7057 test: Enabled test_destroy_invalid_uuid and refactored the tests
test_destroy_invalid_uuid: It was setting invalid UUID to underlying DaosPool,
but we're using dmg, so set the invalid UUID to TestPool instead.

test_destroy_invalid_group: Removed because we're using dmg and there's no
argument to set group for dmg pool destroy. We used to be able to test with API,
but we can't create a pool with API anymore.

Removed unused values in yaml and refactored tags.

Quick-build: true
Skip-unit-tests: true
Test-tag: destroy_single destroy_multi destroy_single_loop destroy_multi_loop destroy_invalid_uuid destroy_connected force_destroy_connected destroy_with_data
Signed-off-by: Makito Kano <makito.kano@intel.com>