DAOS-7066 control: populate system name in control api (#5139)
System name can be configured in the server and dmg config files. When
set for dmg, the control API needs to apply the new system name in
protobuf requests that are issued to the gRPC server. This PR applies
the logic for selecting the system name into the control API and takes
the system name from either the supplied request, the supplied control
config or uses the build default (in that order of preference).

In dmg, make all pool and system subcommands config aware and remove the
ability to set the system name from the commandline. Changing the system
name should instead be performed through the dmg config file which will
work reliably for pool and system commands.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>