DAOS-7192 rebuild: Check if the container is destroyed (#5354)
If enumeration or fetch return -DER_NONEXIST, let's ignore
the failure if the container is destroyed.

For DER_BUSY, let's retry outside scan_request(), so it can
be delay for a few seconds.

Remove incorrect assertion for migrate tls.

Ignore -DER_SHUTDOWN if the container is about to be destoryed
during migration.

Signed-off-by: Di Wang <di.wang@intel.com>