DAOS-7168 object: properly split request for EC object update (#5408)
Usually, for a standalone EC object update, the parity shard with
the largest index will be used as related DTX leader. But for the
distributed transaction, the leader can be any shard of EC object,
or the leader can be some server that out of the EC object layout.
Under such case, when split EC object update request on the leader
via obj_ec_rw_req_split(), we cannot assume that the leader always
be the last shard, instead, the caller needs to explicitly specify
which one is the leader. Otherwise the sub-request to the sent to
other non-leader may be dispatched to the leader server (locally).

Signed-off-by: Fan Yong <fan.yong@intel.com>