Cart-964 CartTest: Resolved cart_iv_two_node valgind error.
Normally, orterun starts shutting down all servers when the first
server is shut down.  Added the --continuous flag to prevent this.
Also, after a server begins to shut down, the server holds off the
complete shutdown for 15 seconds in order to respond to the client
acknowledge.  Otherwise, a crt_context structure is left and
not de-allocated. 15 seconds required by valgrid, 5 seconds normally.

Signed-off-by: Steven Bollinger <stevenx.bollinger@intel.com>